03 Tienda E-commerce (6)/requirements.txt
# eCommerce Modular Backend Dependencies
# Core Flask framework and extensions
Flask==2.3.3
Flask-CORS==4.0.0
Werkzeug==2.3.7

# Database and ORM
SQLAlchemy==2.0.23
PyMySQL==1.1.0
mysqlclient==2.2.0

# Authentication and security
PyJWT==2.8.0
bcrypt==4.1.2
cryptography==41.0.8

# API documentation
flasgger==0.9.7.1

# Data validation and serialization
marshmallow==3.20.1

# Environment variables
python-dotenv==1.0.0

# HTTP client for external APIs
requests==2.31.0

# Date and time utilities
python-dateutil==2.8.2

# Decimal handling
decimal==1.70

# UUID generation
uuid==1.30

# Development and testing
pytest==7.4.3
pytest-flask==1.3.0
pytest-cov==4.1.0

# Code quality
flake8==6.1.0
black==23.11.0

# Production server
gunicorn==21.2.0

# Monitoring and logging
structlog==23.2.0

# Cache (Redis client)
redis==5.0.1

# Background tasks
celery==5.3.4

# Email sending
Flask-Mail==0.9.1

# File upload handling
Pillow==10.1.0

# JSON handling
orjson==3.9.10

# Rate limiting
Flask-Limiter==3.5.0

# Health checks
Flask-HealthCheck==1.0.0

